# controlplane/openapi/axon-control-plane-v1.yaml
openapi: 3.0.3
info:
  title: Axon Control Plane API
  version: 1.0.0
  description: |
    v1 REST surface of the Axon Control Plane (docs/AXON_HOME_ARCHITECTURE.md §5).
    **Status: designed, not built** — no server implementation exists yet; this
    document pins the contract a future in-repo FastAPI service must satisfy.

    Authority: the Control Plane is a CLIENT of the governed seams (Heart /
    Trainer), never a second authority. It reads State-root JSON, dispatches
    `TrainerOrganCommand`s, and calls `HeartHost` methods; it never commits to
    the Shared Field itself. Governed failures (Trainer `UNAVAILABLE`,
    `REJECTED`, stale-delta, sealed-region) pass through verbatim.

    Deployment: binds localhost by default; remote access via SSH tunnel /
    overlay only; never publicly exposed without explicit config + auth.
    TLS is required; bearer tokens are minted during secure device enrolment.

    Error model (§5.6): every error is
    `{schema: "axon-controlplane-error-v1", error: {code, message, retryable, upstream_status?}}`.
    Unknown schema versions are rejected with 426 + UNSUPPORTED_SCHEMA_VERSION
    listing supported versions. All mutating endpoints accept an
    `Idempotency-Key` header; retries never double-apply. Every mutating call
    is audited (§14). Every stateful response carries `freshness`
    (LIVE | CACHED | STALE | DISCONNECTED) and `observed_at`.

    Streaming: `/v1/events` is WebSocket (primary) / SSE (fallback) carrying
...
servers:
...
paths:
...
  /v1/field/head:
    get:
      tags: [field]
      operationId: getFieldHead
      summary: Active branch HEAD + health (HEAD.json + health_latest.json)
      responses:
        '200':
          description: Branch head and health summary
          content:
            application/json:
              schema:
                $ref: '../schemas/field.json#/$defs/fieldHead'
        default:
          $ref: '#/components/responses/Error'
...
  /v1/trainer/status:
    get:
      tags: [trainer]
      operationId: getTrainerStatus
      summary: inspect_trainer_state (axon-trainer-inspection-v1)
      responses:
        '200':
          description: Trainer status
          content:
            application/json:
              schema:
                $ref: '../schemas/trainer.json#/$defs/trainerStatus'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: Bearer token minted during secure device enrolment (§5.3/§8). Required on all endpoints; TLS mandatory.
...
  responses:
...
    Error:
      description: |
        Typed error envelope `{schema, error{code, message, retryable,
        upstream_status?}}`. Governed failures pass through verbatim (§5.6).
        Unknown schema versions → 426 UNSUPPORTED_SCHEMA_VERSION listing
        supported versions.
      content:
        application/json:
          schema:
            $ref: '../schemas/common.json#/$defs/error'
security:
  - bearerAuth: []